+++ src/Sentry/Laravel/Console/TestCommand.php
@@ -11,6 +11,7 @@ use Sentry\State\Hub;
 use Sentry\State\HubInterface;
 use Sentry\Tracing\SpanContext;
 use Sentry\Tracing\TransactionContext;
+use Throwable;
 
 class TestCommand extends Command
 {
@@ -54,9 +55,20 @@ class TestCommand extends Command
 
         $dsn = $this->option('dsn');
 
+        $laravelClient = null;
+
+        try {
+            $laravelClient = app(HubInterface::class)->getClient();
+        } catch (Throwable $e) {
+            // Ignore any errors related to getting the client from the Laravel container
+            // These errors will surface later in the process but we should not crash here
+        }
+
         // If the DSN was not passed as option to the command we use the registered client to get the DSN from the Laravel config
         if ($dsn === null) {
-            $dsnObject = app(HubInterface::class)->getClient()->getOptions()->getDsn();
+            $dsnObject = $laravelClient === null
+                ? null
+                : $laravelClient->getOptions()->getDsn();
 
             if ($dsnObject !== null) {
                 $dsn = (string)$dsnObject;
@@ -77,6 +89,8 @@ class TestCommand extends Command
         try {
             $clientBuilder = ClientBuilder::create([
                 'dsn' => $dsn,
+                'release' => $laravelClient === null ? null : $laravelClient->getOptions()->getRelease(),
+                'environment' => $laravelClient === null ? null : $laravelClient->getOptions()->getEnvironment(),
                 'traces_sample_rate' => 1.0,
             ]);
         } catch (Exception $e) {